refactor: 💡 Unification of types and change of argument data ty
