+++ src/pages/contact/confirm/index.page.tsx
@@ -1,10 +1,6 @@
 import React, { FormEventHandler, useEffect, useState } from 'react';
 import { useRouter } from 'next/router';
 import { UserInput } from '../types';
-import {
-  loadSessionData,
-  removeSessionData,
-} from '@/pages/contact/confirm/logics/loadSession';
 import { postContact } from '../logic/postContact';
 import { formatConfirmField } from '@/pages/contact/confirm/logics/formatConfirmField';
 import { useIsBoolean } from '@/hooks/useBoolean';
@@ -12,6 +8,7 @@ import { createLoadingMessage } from './components/CreateLoadingMessage';
 import { ConfirmPresentation } from './components/ConfirmPresentation';
 import toast from 'react-hot-toast';
 import { TOAST_MESSAGES } from '@/pages/contact/const/message';
+import { loadSessionData, removeSessionData } from './logics/sessionStorage';
 
 export default function Confirm() {
   const [userInput, setUserInput] = useState<UserInput | undefined>();
@@ -19,7 +16,7 @@ export default function Confirm() {
   const router = useRouter();
 
   useEffect(() => {
-    const result = loadSessionData<UserInput>({ label: 'formInput' });
+    const result = loadSessionData<UserInput>('formInput');
     if (result.ok) {
       setUserInput(result.data);
     } else {
